add print_stack_to_json_file
